fix bug: error while binary length larger than short

+++ processing/src/main/java/org/apache/carbondata/processing/loading/sort/SortStepRowHandler.java
@@ -792,7 +792,7 @@ public class SortStepRowHandler implements Serializable {
             reUsableByteArrayDataOutputStream);
       } else {
         byte[] bytes = (byte[]) row[this.noDictNoSortDimIdx[idx]];
-        reUsableByteArrayDataOutputStream.writeShort((short) bytes.length);
+        reUsableByteArrayDataOutputStream.writeInt(bytes.length);
         reUsableByteArrayDataOutputStream.write(bytes);
       }
     }
